refactor: extract comparison state + remove dead storage code
- Create comparison-state.ts: comparisonSet, toggleComparison,
  isInComparison, isComparisonFull, updateCompareBar, announceStatus
- Update imports in rankings-view, city-detail-view, main
- Remove dead driverCount setting and maxTrailers migration
- Remove dead cityTrailers CRUD and related tests
- rankings-view.ts reduced from 672 to ~580 lines

Closes #223
